relecture

+++ libft/basic/ft_strncmp.c
@@ -21,5 +21,5 @@ int	ft_strncmp(const char *s1, const char *s2, size_t n)
 		i++;
 	if (i == n)
 		return (0);
-	return ((unsigned char)s1[i] - (unsigned char)s2[i]);
+	return ((int)(unsigned char)s1[i] - (int)(unsigned char)s2[i]);
 }
